requirements.txt
streamlit
pillow
numpy
torch
opencv-python
